Use url in config payload to set testBaseUrl
- Move traverse functions to ExampleProcessor
