+++ core/src/main/kotlin/io/specmatic/test/ExamplePreProcessor.kt
@@ -7,10 +7,8 @@ import io.specmatic.core.pattern.Row
 import io.specmatic.core.pattern.parsedJSONObject
 import io.specmatic.core.pattern.parsedValue
 import io.specmatic.core.utilities.exceptionCauseMessage
-import io.specmatic.core.value.JSONArrayValue
-import io.specmatic.core.value.JSONObjectValue
-import io.specmatic.core.value.StringValue
-import io.specmatic.core.value.Value
+import io.specmatic.core.value.*
+import io.specmatic.test.asserts.isKeyAssert
 import java.io.File
 
 const val delayedRandomSubstitutionKey = "\$rand"
@@ -36,6 +34,10 @@ object ExampleProcessor {
         return runCatching { parsedJSONObject(configFile.readText()) }.getOrElse { e ->
             consoleLog("Error loading CONFIG $configFilePath: ${exceptionCauseMessage(e)}")
             JSONObjectValue(emptyMap())
+        }.also {
+            it.findFirstChildByPath("url")?.let {
+                url -> System.setProperty("testBaseURL", url.toStringLiteral())
+            }
         }
     }
 
@@ -190,4 +192,38 @@ object ExampleProcessor {
             else -> ""
         }
     }
+}
+
+internal fun <T> Value.traverse(
+    prefix: String = "", onScalar: (Value, String) -> Map<String, T>,
+    onComposite: ((Value, String) -> Map<String, T>)? = null, onAssert: ((Value, String) ->  Map<String, T>)? = null
+): Map<String, T> {
+    return when (this) {
+        is JSONObjectValue -> this.traverse(prefix, onScalar, onComposite, onAssert)
+        is JSONArrayValue ->  this.traverse(prefix, onScalar, onComposite, onAssert)
+        is ScalarValue -> onScalar(this, prefix)
+        else -> emptyMap()
+    }.filterValues { it != null }
+}
+
+private fun <T> JSONObjectValue.traverse(
+    prefix: String = "", onScalar: (Value, String) -> Map<String, T>,
+    onComposite: ((Value, String) -> Map<String, T>)? = null, onAssert: ((Value, String) ->  Map<String, T>)? = null
+): Map<String, T> {
+    return this.jsonObject.entries.flatMap { (key, value) ->
+        val fullKey = if (prefix.isNotEmpty()) "$prefix.$key" else key
+        key.isKeyAssert {
+            onAssert?.invoke(value, fullKey)?.entries.orEmpty()
+        } ?: value.traverse(fullKey, onScalar, onComposite, onAssert).entries
+    }.associate { it.toPair() } + onComposite?.invoke(this, prefix).orEmpty()
+}
+
+private fun <T> JSONArrayValue.traverse(
+    prefix: String = "", onScalar: (Value, String) -> Map<String, T>,
+    onComposite: ((Value, String) -> Map<String, T>)? = null, onAssert: ((Value, String) ->  Map<String, T>)? = null
+): Map<String, T> {
+    return this.list.mapIndexed { index, value ->
+        val fullKey = if (onAssert != null) { prefix } else "$prefix[$index]"
+        value.traverse(fullKey, onScalar, onComposite, onAssert)
+    }.flatMap { it.entries }.associate { it.toPair() } + onComposite?.invoke(this, prefix).orEmpty()
 }
